Janky Kong - June 13th 11:28 - read desc
This code is a combo of everyone's current changes includes the hammer, left-right movement, and the intro. It has been modified to be more efficient and easier to read.

However, there are some issues:

Hammer: It appears, but vanishes when the space bar is pressed.
Left-Right Movement: It fails unpredictably.
Intro: It doesn’t function properly and appears behind the game screen.

We should all download this version and collaborate to resolve these issues.